+++ src/org/linkeddatafragments/servlet/TriplePatternFragmentServlet.java
@@ -20,11 +20,11 @@ import org.apache.jena.riot.RDFDataMgr;
 import org.apache.jena.riot.RDFLanguages;
 import org.linkeddatafragments.config.ConfigReader;
 import org.linkeddatafragments.datasource.DataSourceFactory;
-import org.linkeddatafragments.datasource.HdtDataSourceType;
 import org.linkeddatafragments.datasource.IDataSource;
 import org.linkeddatafragments.datasource.IFragmentRequestProcessor;
-import org.linkeddatafragments.datasource.IndexDataSource;
-import org.linkeddatafragments.datasource.JenaTDBDataSourceType;
+import org.linkeddatafragments.datasource.hdt.HdtDataSourceType;
+import org.linkeddatafragments.datasource.index.IndexDataSource;
+import org.linkeddatafragments.datasource.tdb.JenaTDBDataSourceType;
 import org.linkeddatafragments.exceptions.DataSourceException;
 import org.linkeddatafragments.exceptions.DataSourceNotFoundException;
 import org.linkeddatafragments.fragments.LinkedDataFragment;
